Fixed the display and ball
Fixed a bug with the ball where it was not going to the opponent after the person scored. Additionally fixed the display. Most importantly I made it so that the ball paused and didn't auto continue when the player scored and the ball moved once the player hit the space button